Rename scraper service
